requirements.txt
Django==2.1.7
django-taggit==0.24.0
django-uuslug==1.1.8
Markdown==3.0.1
Pillow==5.4.1
pkg-resources==0.0.0
psycopg2==2.7.7
python-slugify==3.0.0
pytz==2018.9
text-unidecode==1.2
